refactor(): MVC 패턴으로 변경
`Layered Architecture Pattern`으로 구현하고 있던 와중, 해당 아키텍쳐 패턴보다는 `MVC Pattern`이 `자동차 경주 미션`에 더 적합하다고 판단되어 변경함.
