+++ src/main/java/racingcar/controller/RacingGameController.java
@@ -1,27 +1,44 @@
 package racingcar.controller;
 
-import racingcar.service.CarService;
+import racingcar.model.Cars;
+import racingcar.util.validator.InputValidator;
+import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
+import java.util.List;
+
 public class RacingGameController {
-    CarService carService = new CarService();
+    InputView inputView = new InputView();
     OutputView outputView = new OutputView();
+    InputValidator inputValidator = new InputValidator();
+    Cars cars = new Cars();
 
     public void gameStart() {
-        int tryCount = carService.getCarNameAndTryCount();
+        getCarName();
+        int tryCount = getUserTryCount();
         printGameResult(tryCount);
         printGameWinner();
+    }
+
+    private int getUserTryCount() {
+        return inputValidator.checkTryCountInputValidation(inputView.getTryCount());
+    }
+
+    private void getCarName() {
+        List<String> carNameList = cars.splitCarNames(inputView.getCarNames());
+        inputValidator.checkCarNameInputValidation(carNameList);
+        cars.saveCarNameList(carNameList);
     }
 
     private void printGameResult(int tryCount) {
         outputView.printEmptyLine();
         outputView.printGameResultTitle();
         for (int i = 0; i < tryCount; i++) {
-            outputView.printCarPosition(carService.game());
+            outputView.printCarPosition(cars.game());
         }
     }
 
     private void printGameWinner() {
-        outputView.printWinners(carService.getGameWinner());
+        outputView.printWinners(cars.getGameWinner());
     }
 }
