--- src/main/java/racingcar/service/CarService.java
@@ -1,68 +1,0 @@
-package racingcar.service;
-
-import racingcar.domain.Car;
-import racingcar.repository.CarRepository;
-import racingcar.util.RandomNumberGenerator;
-import racingcar.util.validator.InputValidator;
-import racingcar.view.InputView;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-public class CarService {
-    InputView inputView = new InputView();
-    InputValidator inputValidator = new InputValidator();
-    CarRepository carRepository = new CarRepository();
-    RandomNumberGenerator randomNumberGenerator = new RandomNumberGenerator();
-
-    public int getCarNameAndTryCount() {
-        getCarNameAndCheckValidation();
-        return getTryCountAndCheckValidation();
-    }
-
-    private void getCarNameAndCheckValidation() {
-        List<String> carNameList = splitCarNames(inputView.getCarNames());
-        inputValidator.checkCarNameInputValidation(carNameList);
-        saveCarNameList(carNameList);
-    }
-
-    private void saveCarNameList(List<String> carNameList) {
-        for (String carName : carNameList) {
-            carRepository.save(carName);
-        }
-    }
-
-    private int getTryCountAndCheckValidation() {
-        return inputValidator.checkTryCountInputValidation(inputView.getTryCount());
-    }
-
-    private List<String> splitCarNames(String carNames) {
-        return Arrays.stream(carNames.split(",")).toList();
-    }
-
-    public List<Car> game() {
-        for (int i = 0; i < carRepository.carList.size(); i++) {
-            if (randomNumberGenerator.pickRandomNumber() >= 4) {
-                carRepository.moveForward(carRepository.carList.get(i));
-            }
-        }
-        return carRepository.carList;
-    }
-
-    public List<String> getGameWinner() {
-        List<String> winners = new ArrayList<>();
-        int max = 0;
-        for (Car car : carRepository.carList) {
-            if (car.getDistance() == max) {
-                winners.add(car.getName());
-            }
-            if (car.getDistance() > max) {
-                winners.clear();
-                winners.add(car.getName());
-                max = car.getDistance();
-            }
-        }
-        return winners;
-    }
-}
